added deleted category, 80% done with update/edit category, TODO: Save update category into database

+++ WebContent/WEB-INF/jsp/manage-category.jsp
@@ -3,9 +3,11 @@
 
 <header><h1>Manage Category</h1></header>
 
+${info}
+
 <c:forEach items="${categories}" var="category">
 	
-		<h2>${category.name}</h2> <a href="edit-category/${category.id}">Edit</a> <a href="delete-category/${category.id}">Delete</a>
+		<h2>${category.name}</h2> <a href="edit-category/${category.id}.html">Edit</a> <a href="delete-category/${category.id}.html" onclick="return confirm('Are You Sure?')">Delete</a>
 		<p class="info">Description: ${category.description}</p>
 	
 </c:forEach>
